+++ packages/access-control/src/layers/types.ts
@@ -24,7 +24,7 @@ export type AccessControlLayerProps<TPayload extends Record<string, any>=any> = 
 
 export type AccessControlLayer<
   TCollections extends Collections = any,
-  TAccessControl extends AccessControl<TCollections, TAccessControl> = any
+  TAccessControl extends AccessControl<any, TAccessControl> = any
 > = (context: Context<any, TCollections>, props: AccessControlLayerProps) => Promise<Either<
   ACErrors,
   ReadPayload | WritePayload