+++ packages/access-control/src/types.ts
@@ -1,48 +1,38 @@
-import type { Collections } from '@sonata-api/api'
+import type { CollectionStructure } from '@sonata-api/api'
 import type { AccessControlLayer } from './layers/types'
 export { AccessControlLayer }
 
 // #region Role
 export type Role<
-  TCollections extends Collections=any,
-  TAccessControl extends AccessControl<TCollections>=any
+  TCollectionStructure extends CollectionStructure=any,
+  TAccessControl extends AccessControl<TCollectionStructure> = any
 > = {
   inherit?: Array<keyof TAccessControl['roles']>
   grantEverything?: boolean
   forbidEverything?: boolean
-  capabilities?: {
-    [P in keyof TCollections]?: {
-      grantEverything?: boolean
-      forbidEverything?: boolean
-      functions?: 'functions' extends keyof TCollections[P]
-        ? Array<keyof TCollections[P]['functions']>
-        : never
-      blacklist?: 'functions' extends keyof TCollections[P]
-        ? Array<keyof TCollections[P]['functions']>
-        : never
-    }
-  }
+  functions?: Array<keyof TCollectionStructure['functions']>
+  blacklist?: Array<keyof TCollectionStructure['functions']>
 }
 // #endregion Role
 
 export type Roles<
-  TCollections extends Collections=any,
-  TAccessControl extends AccessControl<TCollections>=any
-> = Record<string, Role<TCollections, TAccessControl>>
+  TCollectionStructure extends CollectionStructure=any,
+  TAccessControl extends AccessControl<TCollectionStructure> = any
+> = Record<string, Role<TCollectionStructure, TAccessControl>>
 
 // #region AccessControl
 export type InternalAccessControl<
-  TCollections extends Collections,
-  TAccessControl extends AccessControl<TCollections>=any
+  TCollectionStructure extends CollectionStructure=any,
+  TAccessControl extends AccessControl<TCollectionStructure> = any
 > = {
-  roles?: Roles<TCollections, TAccessControl>
+  roles?: Roles<TCollectionStructure, TAccessControl>
   availableRoles?: keyof TAccessControl['roles']
   parent?: TAccessControl['roles']
 }
 
 export type AccessControl<
-  TCollections extends Collections=any,
-  TAccessControl extends AccessControl<TCollections, TAccessControl>=any
-> = InternalAccessControl<TCollections, TAccessControl>
+  TCollectionStructure extends CollectionStructure=any,
+  TAccessControl extends AccessControl<TCollectionStructure, TAccessControl> = any
+> = InternalAccessControl<TCollectionStructure, TAccessControl>
 // #endregion AccessControl