+++ packages/system/src/collections/user/index.ts
@@ -1,5 +1,4 @@
 import { defineCollection, useFunctions } from '@sonata-api/api'
-import { defineAccessControl } from '@sonata-api/access-control'
 import { description, User } from './description'
 
 import authenticate from './authenticate'
@@ -25,15 +24,13 @@ export const user = defineCollection(() => ({
     getInfo,
     activate,
   },
-  // accessControl: defineAccessControl()({
-  //   roles: {
-  //     guest: {
-  //       capabilities: {
-  //         functions: [
-  //           'authenticate'
-  //         ]
-  //       }
-  //     }
-  //   }
-  // })
+  accessControl: {
+    roles: {
+      guest: {
+        functions: [
+          'authenticate'
+        ]
+      }
+    }
+  }
 }))